(TextAtlas) Let Text relative geometry use uint32_t indices
Due to the data format is uint16_t type, we cannot render well
if the length of text is bigger than 16384.
If we add underline or strokethrough, the number of text limitation
also be reduced.

To fix this issue, let we make indices what TextAtlas geometry use
uint32_t type, instead uint16_t type.

Since we support uint32_t* indices now a day, it will be okay.

+++ dali-toolkit/internal/text/decorator/text-decorator.cpp
@@ -1219,8 +1219,8 @@ struct Decorator::Impl : public ConnectionTracker
         const float offsetX = mHighlightPosition.x + 0.5f * mHighlightSize.width;
         const float offsetY = mHighlightPosition.y + 0.5f * mHighlightSize.height;
 
-        Vector<Vector2>        vertices;
-        Vector<unsigned short> indices;
+        Vector<Vector2>  vertices;
+        Vector<uint32_t> indices;
 
         vertices.Reserve(4u * numberOfQuads);
         indices.Reserve(6u * numberOfQuads);
